WIP: layout for 'all' apps WIP: setting up viewmodel and activity's functionalities Added event bus Added network error validation and layout

+++ app/src/main/java/com/example/miketoide/data/persistence/App.kt
@@ -10,7 +10,7 @@ data class App(
     val id : String,
     val name: String,
     val version: String,
-    val downloads: Int,
+    val downloads: Long,
     val launchDate: String,
     val rating: Double,
     val icon: String,
